+++ src/AlexaSchema.ts
@@ -19,11 +19,26 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-/* tslint:disable:no-empty */
+/* tslint:disable:no-empty no-submodule-imports */
 import * as _Promise from "bluebird";
 import * as _ from "lodash";
 import * as path from "path";
-import { IIntent, Schema } from "./Schema";
+import * as uuid from "uuid/v5";
+import {
+  DelegationStrategy,
+  IDialog,
+  IDialogIntent,
+  IDialogSlot,
+  IDialogSlotPrompt,
+  IInteractionModel,
+  IInteractionModelIntent,
+  IInteractionModelSlot,
+  IInteractionModelType,
+  IInteractionModelTypeValue,
+  IPrompt,
+  IVariation
+} from "./AlexaInterfaces";
+import { IIntent, ISlotDefinition, Schema } from "./Schema";
 import { IVoxaSheet } from "./VoxaSheet";
 
 const NAMESPACE = "alexa";
@@ -72,7 +87,7 @@ export class AlexaSchema extends Schema {
       content: { manifest }
     });
   }
-  public contentLanguageModel(locale: string, environment: string) {
+  public contentLanguageModel(locale: string, environment: string): IInteractionModel {
     const invocation = _.find(this.invocations, { locale, environment });
     const invocationName = _.get(invocation, "name", "Skill with no name");
 
@@ -83,7 +98,22 @@ export class AlexaSchema extends Schema {
       return slot;
     });
 
-    return { interactionModel: { languageModel: { invocationName, intents, types } } };
+    const dialog: IDialog = {
+      intents: this.generateDialogModel(this.intentsByPlatformAndEnvironments(locale, environment)),
+      delegationStrategy: "SKILL_RESPONSE"
+    };
+
+    const prompts: IPrompt[] = this.generatePrompts(
+      this.intentsByPlatformAndEnvironments(locale, environment)
+    );
+
+    return {
+      interactionModel: {
+        languageModel: { invocationName, intents, types },
+        dialog,
+        prompts
+      }
+    };
   }
   public buildLanguageModel(locale: string, environment: string) {
     this.fileContent.push({
@@ -111,4 +141,140 @@ export class AlexaSchema extends Schema {
       content: canFulfillIntents
     });
   }
+
+  private generateDialogModel(intents: IIntent[]): IDialogIntent[] {
+    return _(intents)
+      .filter(intentOrSlotRequireDialog)
+      .map(
+        (intent): IDialogIntent => {
+          const prompts: IDialogSlotPrompt = {};
+          if (intent.confirmations.length > 0) {
+            prompts.confirmation = getPromptId("Confirmation", "Intent", intent.confirmations);
+          }
+          return {
+            name: intent.name,
+            delegationStrategy: intent.delegationStrategy,
+            confirmationRequired: intent.confirmationRequired,
+            slots: this.generateDialogSlotModel(intent.slotsDefinition),
+            prompts
+          };
+        }
+      )
+      .value();
+  }
+
+  private generateDialogSlotModel(slots: ISlotDefinition[]): IDialogSlot[] {
+    return _(slots)
+      .map(
+        (slot: ISlotDefinition): IDialogSlot => {
+          const prompts: IDialogSlotPrompt = {};
+          if (slot.prompts.elicitation.length > 0) {
+            prompts.elicitation = getPromptId("Elicitation", "Slot", slot.prompts.elicitation);
+          }
+
+          if (slot.prompts.confirmation.length > 0) {
+            prompts.confirmation = getPromptId("Confirmation", "Slot", slot.prompts.confirmation);
+          }
+
+          return {
+            name: slot.name.replace("{", "").replace("}", ""),
+            type: slot.type,
+            elicitationRequired: slot.requiresElicitation,
+            confirmationRequired: slot.requiresConfirmation,
+            prompts
+          };
+        }
+      )
+      .value();
+  }
+
+  private generatePrompts(intents: IIntent[]): IPrompt[] {
+    const intentPrompts: IPrompt[] = _(intents)
+      .filter(intentHasPrompts)
+      .map(
+        (intent: IIntent): IPrompt =>
+          getPromptsObject("Confirmation", "Intent", intent.confirmations)
+      )
+      .value();
+
+    const slotPrompts = this.generateSlotPrompts(intents);
+
+    return _.concat(intentPrompts, slotPrompts);
+  }
+
+  private generateSlotPrompts(intents: IIntent[]): IPrompt[] {
+    return _(intents)
+      .map("slotsDefinition")
+      .flatten()
+      .filter(slotHasPrompts)
+      .map(
+        (slot: ISlotDefinition): IPrompt[] => {
+          const prompts: IPrompt[] = [];
+          if (slot.prompts.confirmation.length > 0) {
+            prompts.push(getPromptsObject("Confirmation", "Slot", slot.prompts.confirmation));
+          }
+
+          if (slot.prompts.elicitation.length > 0) {
+            prompts.push(getPromptsObject("Elicitation", "Slot", slot.prompts.elicitation));
+          }
+
+          return prompts;
+        }
+      )
+      .flatten()
+      .value();
+  }
+}
+
+function hashObj(obj: any): string {
+  return uuid(JSON.stringify(obj), uuid.DNS);
+}
+
+function intentOrSlotRequireDialog(intent: IIntent): boolean {
+  if (intent.confirmationRequired || intent.delegationStrategy) {
+    return true;
+  }
+
+  const slotRequiresDialog = _(intent.slotsDefinition)
+    .map(slot => slot.requiresElicitation || slot.requiresConfirmation)
+    .some();
+
+  return slotRequiresDialog;
+}
+
+function slotHasPrompts(slot: ISlotDefinition): boolean {
+  return slot.prompts.confirmation.length > 0 || slot.prompts.elicitation.length > 0;
+}
+
+function intentHasPrompts(intent: IIntent): boolean {
+  return intent.confirmations.length > 0;
+}
+
+function getPromptsObject(
+  dialogType: "Elicitation" | "Confirmation",
+  objectType: "Slot" | "Intent",
+  data: string[]
+): IPrompt {
+  return {
+    id: getPromptId(dialogType, objectType, data),
+    variations: formatVariations(data)
+  };
+}
+
+function getPromptId(
+  dialogType: "Elicitation" | "Confirmation",
+  objectType: "Slot" | "Intent",
+  data: string[]
+): string {
+  return `${dialogType}.${objectType}.${hashObj(data)}`;
+}
+
+function formatVariations(variations: string[]): IVariation[] {
+  return _.map(
+    variations,
+    (variation): IVariation => ({
+      type: "PlainText",
+      value: variation
+    })
+  );
 }