+++ templates/javascript/src/app/states/index.js
@@ -1,0 +1,7 @@
+const mainStates = require("./main.states");
+
+function register(voxaApp) {
+  mainStates(voxaApp);
+}
+
+module.exports = register;