+++ templates/javascript/src/app/variables.js
@@ -1,0 +1,3 @@
+exports.name = function(voxaEvent) {
+  return voxaEvent.model.name;
+};