+++ templates/javascript/web/handler.js
@@ -1,0 +1,4 @@
+const serverless = require("serverless-http");
+const { expressApp } = require("../server");
+
+exports.expressHandler = serverless(expressApp);